refactor(core): extract standby-store + standby-scheduler (KJC-TSK-0511 PR8)
Mueve src/brain/standby-store.js y src/brain/standby-scheduler.js a
@karajan/core/standby-store y @karajan/core/standby-scheduler. Los
módulos antiguos pasan a ser shims de re-export para mantener la API
del CLI estable.

Con este PR @karajan/hu-board no tiene NINGÚN import (estático ni
dinámico) hacia ../../../src/. PR9 podrá re-añadir packages/hu-board
al array `workspaces` raíz.

- packages/core/src/standby-store.js (verbatim, imports recableados a ./)
- packages/core/src/standby-scheduler.js (verbatim, import a ./standby-store)
- packages/core/package.json: 2 nuevos subpath exports
- packages/core/src/index.js: 2 nuevos export *
- packages/core/tests/standby.test.js: 5 tests (persist+load+list, done,
  lock exclusivo, scheduleResume immediate-on-expired, cancelScheduled)
- packages/core/README.md: 2 filas + nota de status PR8
- src/brain/standby-{store,scheduler}.js: shims re-export
- packages/hu-board/src/server.js: dynamic import a @karajan/core/standby-scheduler
- packages/hu-board/tests/api.test.js: dynamic import a @karajan/core/standby-store

Tests:
- packages/core 5/5 verde
- CLI shims tests/standby.test.js + tests/brain/standby-*.test.js 32/32
- hu-board tests/api.test.js 29/29
