+++ packages/core/src/index.js
@@ -14,3 +14,5 @@ export * from "./plan-hu-ops.js";
 export * from "./plan-validation.js";
 export * from "./process.js";
 export * from "./hu-snapshot.js";
+export * from "./standby-store.js";
+export * from "./standby-scheduler.js";
